// packages/protocol/typescript/generated/agent_start/runtime/v1/session_tabs_pb.d.ts
export declare type SessionTabsPaneLayoutNode = Message<"agent_start.runtime.v1.SessionTabsPaneLayoutNode"> & {
  /**
   * @generated from oneof agent_start.runtime.v1.SessionTabsPaneLayoutNode.node
   */
  node: {
    /**
     * @generated from field: agent_start.runtime.v1.SessionTabsPaneLeaf leaf = 1;
     */
    value: SessionTabsPaneLeaf;
    case: "leaf";
  } | {
    /**
     * @generated from field: agent_start.runtime.v1.SessionTabsPaneSplit split = 2;
     */
    value: SessionTabsPaneSplit;
    case: "split";
  } | { case: undefined; value?: undefined };
};

/**
 * Describes the message agent_start.runtime.v1.SessionTabsPaneLayoutNode.
 * Use `create(SessionTabsPaneLayoutNodeSchema)` to create a new message.
 */
export declare const SessionTabsPaneLayoutNodeSchema: GenMessage<SessionTabsPaneLayoutNode>;

/**
 * @generated from message agent_start.runtime.v1.SessionTabsPaneLeaf
 */
export declare type SessionTabsPaneLeaf = Message<"agent_start.runtime.v1.SessionTabsPaneLeaf"> & {
  /**
   * @generated from field: string leaf_id = 1;
   */
  leafId: string;
};

/**
 * Describes the message agent_start.runtime.v1.SessionTabsPaneLeaf.
 * Use `create(SessionTabsPaneLeafSchema)` to create a new message.
 */
export declare const SessionTabsPaneLeafSchema: GenMessage<SessionTabsPaneLeaf>;

/**
 * @generated from message agent_start.runtime.v1.SessionTabsPaneSplit
 */
export declare type SessionTabsPaneSplit = Message<"agent_start.runtime.v1.SessionTabsPaneSplit"> & {
  /**
   * @generated from field: agent_start.runtime.v1.SessionTabsPaneSplitDirection direction = 1;
   */
  direction: SessionTabsPaneSplitDirection;

  /**
   * @generated from field: agent_start.runtime.v1.SessionTabsPaneLayoutNode first = 2;
   */
  first?: SessionTabsPaneLayoutNode | undefined;

  /**
   * @generated from field: agent_start.runtime.v1.SessionTabsPaneLayoutNode second = 3;
   */
  second?: SessionTabsPaneLayoutNode | undefined;

  /**
   * @generated from field: optional double ratio = 4;
   */
  ratio?: number | undefined;
};

/**
 * Describes the message agent_start.runtime.v1.SessionTabsPaneSplit.
 * Use `create(SessionTabsPaneSplitSchema)` to create a new message.
 */
export declare const SessionTabsPaneSplitSchema: GenMessage<SessionTabsPaneSplit>;
// ...
export declare type SessionTabsNullableColor = Message<"agent_start.runtime.v1.SessionTabsNullableColor"> & {
  /**
   * @generated from oneof agent_start.runtime.v1.SessionTabsNullableColor.value
   */
  value: {
    /**
     * @generated from field: string text = 1;
     */
    value: string;
    case: "text";
  } | {
    /**
     * @generated from field: bool null = 2;
     */
    value: boolean;
    case: "null";
  } | { case: undefined; value?: undefined };
};

/**
 * Describes the message agent_start.runtime.v1.SessionTabsNullableColor.
 * Use `create(SessionTabsNullableColorSchema)` to create a new message.
 */
export declare const SessionTabsNullableColorSchema: GenMessage<SessionTabsNullableColor>;

/**
 * @generated from message agent_start.runtime.v1.SessionTabsLaunchConfig
 */
export declare type SessionTabsLaunchConfig = Message<"agent_start.runtime.v1.SessionTabsLaunchConfig"> & {
  /**
   * @generated from field: string agent_args = 1;
   */
  agentArgs: string;

  /**
   * @generated from field: map<string, string> agent_env = 2;
   */
  agentEnv: { [key: string]: string };

  /**
   * @generated from field: optional string agent_command = 3;
   */
  agentCommand?: string | undefined;

  /**
   * @generated from field: optional string omp_resume_file_path = 4;
   */
  ompResumeFilePath?: string | undefined;
};

/**
 * Describes the message agent_start.runtime.v1.SessionTabsLaunchConfig.
 * Use `create(SessionTabsLaunchConfigSchema)` to create a new message.
 */
export declare const SessionTabsLaunchConfigSchema: GenMessage<SessionTabsLaunchConfig>;

/**
 * @generated from message agent_start.runtime.v1.SessionTabsStartupCommandDelivery
 */
export declare type SessionTabsStartupCommandDelivery = Message<"agent_start.runtime.v1.SessionTabsStartupCommandDelivery"> & {
  /**
   * @generated from oneof agent_start.runtime.v1.SessionTabsStartupCommandDelivery.delivery
   */
  delivery: {
    /**
     * @generated from field: bool fast = 1;
     */
    value: boolean;
    case: "fast";
  } | {
    /**
     * @generated from field: bool shell_ready = 2;
     */
    value: boolean;
    case: "shellReady";
  } | { case: undefined; value?: undefined };
};

/**
 * Describes the message agent_start.runtime.v1.SessionTabsStartupCommandDelivery.
 * Use `create(SessionTabsStartupCommandDeliverySchema)` to create a new message.
 */
export declare const SessionTabsStartupCommandDeliverySchema: GenMessage<SessionTabsStartupCommandDelivery>;
// ...
export declare type SessionTabsServiceMoveReorder = Message<"agent_start.runtime.v1.SessionTabsServiceMoveReorder"> & {
  /**
   * @generated from field: repeated string tab_order = 1;
   */
  tabOrder: string[];
};

/**
 * Describes the message agent_start.runtime.v1.SessionTabsServiceMoveReorder.
 * Use `create(SessionTabsServiceMoveReorderSchema)` to create a new message.
 */
export declare const SessionTabsServiceMoveReorderSchema: GenMessage<SessionTabsServiceMoveReorder>;

/**
 * @generated from message agent_start.runtime.v1.SessionTabsServiceMoveToGroup
 */
export declare type SessionTabsServiceMoveToGroup = Message<"agent_start.runtime.v1.SessionTabsServiceMoveToGroup"> & {
  /**
   * @generated from field: optional uint32 index = 1;
   */
  index?: number | undefined;
};

/**
 * Describes the message agent_start.runtime.v1.SessionTabsServiceMoveToGroup.
 * Use `create(SessionTabsServiceMoveToGroupSchema)` to create a new message.
 */
export declare const SessionTabsServiceMoveToGroupSchema: GenMessage<SessionTabsServiceMoveToGroup>;

/**
 * @generated from message agent_start.runtime.v1.SessionTabsServiceMoveSplit
 */
export declare type SessionTabsServiceMoveSplit = Message<"agent_start.runtime.v1.SessionTabsServiceMoveSplit"> & {
  /**
   * @generated from field: agent_start.runtime.v1.SessionTabsSplitDirection direction = 1;
   */
  direction: SessionTabsSplitDirection;
};

/**
 * Describes the message agent_start.runtime.v1.SessionTabsServiceMoveSplit.
 * Use `create(SessionTabsServiceMoveSplitSchema)` to create a new message.
 */
export declare const SessionTabsServiceMoveSplitSchema: GenMessage<SessionTabsServiceMoveSplit>;
// ...
export declare type SessionTabsPaneLayoutRoot = Message<"agent_start.runtime.v1.SessionTabsPaneLayoutRoot"> & {
  /**
   * @generated from oneof agent_start.runtime.v1.SessionTabsPaneLayoutRoot.value
   */
  value: {
    /**
     * @generated from field: bool null = 1;
     */
    value: boolean;
    case: "null";
  } | {
    /**
     * @generated from field: agent_start.runtime.v1.SessionTabsPaneLayoutNode node = 2;
     */
    value: SessionTabsPaneLayoutNode;
    case: "node";
  } | { case: undefined; value?: undefined };
};

/**
 * Describes the message agent_start.runtime.v1.SessionTabsPaneLayoutRoot.
 * Use `create(SessionTabsPaneLayoutRootSchema)` to create a new message.
 */
export declare const SessionTabsPaneLayoutRootSchema: GenMessage<SessionTabsPaneLayoutRoot>;

/**
 * @generated from message agent_start.runtime.v1.SessionTabsNullableStringField
 */
export declare type SessionTabsNullableStringField = Message<"agent_start.runtime.v1.SessionTabsNullableStringField"> & {
  /**
   * @generated from oneof agent_start.runtime.v1.SessionTabsNullableStringField.value
   */
  value: {
    /**
     * @generated from field: string text = 1;
     */
    value: string;
    case: "text";
  } | {
    /**
     * @generated from field: bool null = 2;
     */
    value: boolean;
    case: "null";
  } | { case: undefined; value?: undefined };
};

/**
 * Describes the message agent_start.runtime.v1.SessionTabsNullableStringField.
 * Use `create(SessionTabsNullableStringFieldSchema)` to create a new message.
 */
export declare const SessionTabsNullableStringFieldSchema: GenMessage<SessionTabsNullableStringField>;
// ...
export enum SessionTabsMoveKind {
  /**
   * @generated from enum value: SESSION_TABS_MOVE_KIND_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: SESSION_TABS_MOVE_KIND_REORDER = 1;
   */
  REORDER = 1,

  /**
   * @generated from enum value: SESSION_TABS_MOVE_KIND_MOVE_TO_GROUP = 2;
   */
  MOVE_TO_GROUP = 2,

  /**
   * @generated from enum value: SESSION_TABS_MOVE_KIND_SPLIT = 3;
   */
  SPLIT = 3,
}

/**
 * Describes the enum agent_start.runtime.v1.SessionTabsMoveKind.
 */
export declare const SessionTabsMoveKindSchema: GenEnum<SessionTabsMoveKind>;

/**
 * @generated from enum agent_start.runtime.v1.SessionTabsSplitDirection
 */
export enum SessionTabsSplitDirection {
  /**
   * @generated from enum value: SESSION_TABS_SPLIT_DIRECTION_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: SESSION_TABS_SPLIT_DIRECTION_LEFT = 1;
   */
  LEFT = 1,

  /**
   * @generated from enum value: SESSION_TABS_SPLIT_DIRECTION_RIGHT = 2;
   */
  RIGHT = 2,

  /**
   * @generated from enum value: SESSION_TABS_SPLIT_DIRECTION_UP = 3;
   */
  UP = 3,

  /**
   * @generated from enum value: SESSION_TABS_SPLIT_DIRECTION_DOWN = 4;
   */
  DOWN = 4,
}

/**
 * Describes the enum agent_start.runtime.v1.SessionTabsSplitDirection.
 */
export declare const SessionTabsSplitDirectionSchema: GenEnum<SessionTabsSplitDirection>;

/**
 * @generated from enum agent_start.runtime.v1.SessionTabsPaneSplitDirection
 */
export enum SessionTabsPaneSplitDirection {
  /**
   * @generated from enum value: SESSION_TABS_PANE_SPLIT_DIRECTION_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: SESSION_TABS_PANE_SPLIT_DIRECTION_HORIZONTAL = 1;
   */
  HORIZONTAL = 1,

  /**
   * @generated from enum value: SESSION_TABS_PANE_SPLIT_DIRECTION_VERTICAL = 2;
   */
  VERTICAL = 2,
}